@@ -8,18 +8,21 @@ package consumer
 import (
 	"fmt"
 	"log"
+	"slices"
 
+	"github.com/kirill-scherba/command/v2"
 	"github.com/teonet-go/teomq"
-	"github.com/teonet-go/teomq/commands"
 	"github.com/teonet-go/teonet"
 )
+
+const logprefix = "consumer: "
 
 // Consumer is Teonet messages queue consumer type.
 type Consumer struct {
 	*teonet.Teonet
 	*teonet.APIClient
 	ProcessMessage
-	*commands.Commands
+	*command.Commands
 }
 
 type ProcessMessage func(p *teonet.Packet) (answer []byte, err error)
@@ -40,7 +43,7 @@ type API bool
 //
 //	*Consumer: new Teonet MQueue Consumer object
 //	error: error if occurred
-func New(appShort, broker string, reader ProcessMessage, attr ...interface{}) (
+func New(appShort, broker string, reader ProcessMessage, attr ...any) (
 	co *Consumer, err error) {
 
 	// Create new consumer object and connect to teonet
@@ -72,8 +75,10 @@ func New(appShort, broker string, reader ProcessMessage, attr ...interface{}) (
 				co.API(broker)
 			}
 
-			// Subscribe to broker commands
-			err = co.subscribeCommands(broker)
+			// Subscribe to broker commands in command mode
+			if co.Commands != nil {
+				err = co.subscribeCommands(broker)
+			}
 		}()
 	})
 
@@ -86,42 +91,65 @@ func New(appShort, broker string, reader ProcessMessage, attr ...interface{}) (
 	return
 }
 
-// API connects to brokers api
+// API connects to brokers api.
 func (co *Consumer) API(broker string) (err error) {
 
 	co.APIClient, err = co.Teonet.NewAPIClient(broker)
 	if err != nil {
-		log.Println("can't connect to broker api, error:", err)
+		log.Println(logprefix+"can't connect to broker api, error:", err)
 		return
 	}
-	log.Println("connected to broker api:", co.APIClient.String())
+	log.Println(logprefix+"connected to broker api:", co.APIClient.String())
 
 	return
 }
 
 // subscribe subscribe to brokers command.
+//
+// Args:
+//
+//	broker: broker address
+//	command: command name to subscribe
+//
+// Returns:
+//
+//	error: error if occurred
 func (co *Consumer) subscribe(broker, command string) (err error) {
-	data := []byte(fmt.Sprintf("subscribe/%s", command))
+	// Send subscribe command to broker
+	data := fmt.Appendf(nil, "subscribe/%s", command)
+
 	if co.APIClient == nil {
+		// Send to broker directly
 		co.Teonet.SendTo(broker, data)
 		return
 	}
+	// Send to broker using API
 	co.APIClient.SendTo("msg", data)
 	return
 }
 
 // addCommands adds command schema to consumer.
-func (co *Consumer) addCommands(attr ...interface{}) (outattr []interface{}) {
+//
+// If function from command package is found in attributes list, it is removed
+// from list and command schema is set to consumer.
+//
+// Args:
+//
+//	attr: teonet application attributes
+//
+// Returns:
+//
+//	outattr: attributes list without command schema
+func (co *Consumer) addCommands(attr ...any) (outattr []any) {
 
 	outattr = attr
 	for i, v := range attr {
 		switch v := v.(type) {
-		case func(*commands.Commands):
-			fmt.Println("Command schema is on")
-			outattr = append(attr[:i], attr[i+1:]...)
+		case func(*command.Commands):
+			log.Println(logprefix + "command schema is on")
+			outattr = slices.Delete(outattr, i, i+1)
 
-			co.Commands = new(commands.Commands)
-			co.Commands.Init()
+			co.Commands = command.New()
 
 			v(co.Commands)
 			return
@@ -131,13 +159,26 @@ func (co *Consumer) addCommands(attr ...interface{}) (outattr []interface{}) {
 	return
 }
 
-func (co *Consumer) addAPI(attr ...interface{}) (outattr []interface{}, ok bool) {
+// addAPI adds API flag to consumer.
+//
+// If API flag is found in attributes list, it is removed from list and
+// flag is set to true.
+//
+// Args:
+//
+//	attr: teonet application attributes
+//
+// Returns:
+//
+//	outattr: teonet application attributes without API flag
+//	ok: true if API flag is found in attributes
+func (co *Consumer) addAPI(attr ...any) (outattr []any, ok bool) {
 	outattr = attr
 	for i, v := range attr {
 		switch v.(type) {
 		case API:
 			fmt.Println("API is on")
-			outattr = append(attr[:i], attr[i+1:]...)
+			outattr = slices.Delete(outattr, i, i+1)
 			ok = true
 			return
 		}
@@ -147,9 +188,9 @@ func (co *Consumer) addAPI(attr ...interface{}) (outattr []interface{}, ok bool)
 
 // subscribeCommands subscribe to brokers commands.
 func (co *Consumer) subscribeCommands(broker string) (err error) {
-	co.Commands.ForEach(func(command string, cmd *commands.CommandData) {
+	for command := range co.Iter() {
 		err = co.subscribe(broker, command)
-	})
+	}
 	return
 }
 
@@ -174,14 +215,14 @@ func (co *Consumer) reader(c *teonet.Channel, p *teonet.Packet,
 
 	// On connected
 	if e.Event == teonet.EventConnected {
-		log.Printf("connected to %s\n", c)
+		log.Printf(logprefix+"connected to %s\n", c)
 		c.Send(teomq.ConsumerHello)
 		return false
 	}
 
 	// On disconnected
 	if e.Event == teonet.EventDisconnected {
-		log.Printf("disconnected from %s\n", c)
+		log.Printf(logprefix+"disconnected from %s\n", c)
 		return false
 	}
 
@@ -202,7 +243,7 @@ func (co *Consumer) reader(c *teonet.Channel, p *teonet.Packet,
 		// Check consumerHello message from new consumer
 		if len(p.Data()) == len(teomq.ConsumerAnswer) &&
 			string(p.Data()) == string(teomq.ConsumerAnswer) {
-			log.Printf("connected to broker\n")
+			log.Printf(logprefix + "connected to broker\n")
 			return true
 		}
 
@@ -216,25 +257,37 @@ func (co *Consumer) reader(c *teonet.Channel, p *teonet.Packet,
 
 			// Execute command
 			case co.Commands != nil:
-				var parts []string
-				_, parts, _, err = co.Commands.Unmarshal(p.Data())
-				if err == nil {
-					answer, err = co.Commands.Exec(parts[0], commands.Teonet, p.Data())
+				// Parse command
+				_, name, vars, data, err := co.ParseCommand(p.Data())
+				if err != nil {
+					log.Printf("parse message id %d, from %s, error: %s\n",
+						p.ID(), c, err)
+					return
+				}
+
+				// Execute command using default request
+				answer, err = co.Commands.Exec(name, command.Teonet,
+					&command.DefaultRequest{Vars: vars, Data: data},
+				)
+				if err != nil {
+					log.Printf(logprefix+"execute command %s, id %d, from %s, error: %s\n",
+						name, p.ID(), c, err)
+					return
 				}
 
 			// Execute custom reader
 			case co.ProcessMessage != nil:
 				answer, err = co.ProcessMessage(p)
+				if err != nil {
+					log.Printf(logprefix+"process message in custom reader eith id %d, "+
+						"from %s, error: %s\n",
+						p.ID(), c, err)
+					return
+				}
 
 			// Default answer if commands and reader does not added
 			default:
 				answer = []byte("Answer to " + string(p.Data()))
-			}
-
-			// Check error
-			if err != nil {
-				log.Printf("process message id %d, from %s, error: %s\n", p.ID(), c, err)
-				return
 			}
 
 			// Don't send empty answer
@@ -245,9 +298,11 @@ func (co *Consumer) reader(c *teonet.Channel, p *teonet.Packet,
 			// Send answer
 			err = co.sendAnswer(p, answer)
 			if err != nil {
-				log.Printf("send id %d, len: %d, to %s, error: %s\n", p.ID(), len(answer), c, err)
+				log.Printf(logprefix+"send id %d, len: %d, to %s, error: %s\n",
+					p.ID(), len(answer), c, err)
 			} else {
-				// log.Printf("send id %d, len: %d, to %s\n", p.ID(), len(answer), c)
+				// log.Printf("send id %d, len: %d, to %s\n",
+				//	p.ID(), len(answer), c)
 			}
 		}()
 
